Refactor sync triggers to own hook.
